Remove logstash_commons dependency
Migrated the `logstash_commons` calls to use attributes on elkstack instead. Added tests to verify the templates are pushed out successfully when converged with elkstack::agent in a wrapper recipe. If elkstack isn't present, these additional attributes simply do nothing. If elkstack is present but a particular technology is not present, logstash simply ignores log files it doesn't find.